Controller atualizado e mudança na conexão com banco de dados de persistence para configuration
