+++ src/main.php
@@ -10,7 +10,7 @@ include "./src/model/Usuario.php";
 include "./src/persistence/ConnectionFactory.php";
 
 use MeuProjeto\model\Login;
-use MeuProjeto\persistence\ConnectionFactory;
+use MeuProjeto\configuration\ConnectionFactory;
 
 // Cria uma instância da classe Login
 $login = new Login($arg1, $arg2); // Replace $arg1 and $arg2 with the actual arguments required by the Login constructor
